Fix for motion_player.cpp and accumulator_set.hpp
- Fixed a bug in motion_player.cpp where the loop counter was of the wrong type.
- Fixed an assignment operator issue in accumulator_set.hpp.

+++ src/motion_player.cpp
@@ -94,7 +94,7 @@ TypedArray<Dictionary> MotionPlayer::query_pose(int64_t included_category, int64
 		TypedArray<Dictionary> results = {};
 
 		String debug = "[";
-		for (int32_t i = 0; i < re.size(); ++i) {
+		for (size_t i = 0; i < re.size(); ++i) {
 			Dictionary subresult{};
 			List<StringName> names;
 			animation_library->get_animation_list(&names);
